+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Tests for "Blowing fuse".
+- Tests for "Container terminal".
 
 ## [1.4.0] - 2022-03-03
 ### Added